+++ src/data/actions/create_cart_action.ts
@@ -1,0 +1,6 @@
+'use server';
+
+export async function createCartAction(
+  _: { status?: boolean; error: string },
+  formData: FormData,
+) {}
